make maxmem an int

+++ include/input.h
@@ -25,7 +25,7 @@ public:
 struct UserInputKreeq : UserInput {
 
     uint8_t kmerLen = 21, covCutOff = 0, depth = 3, backtrackingSpan = 5;
-    double maxMem = 0;
+    uint64_t maxMem = 0;
     std::string prefix = ".", outFile = "";
     std::vector<std::string> inDBG;
